refactor: Move bufferCrankOutput and flushCrankBuffer to crank.ts
